+++ src/components/ProjectInput/index.js
@@ -5,7 +5,7 @@ import { connect } from 'react-redux';
 
 import Dropdown from 'semantic-ui-react/dist/commonjs/modules/Dropdown';
 
-import { sortedProjects } from '../../selectors/projects';
+import { sortedProjects } from 'sections/Projects/selectors';
 
 type ProjectInputType = {
   placeholder?: string,
