+++ src/sections/Projects/selectors.js
@@ -2,7 +2,7 @@
 
 import { createSelector } from 'reselect';
 
-import { sortProjects } from '../core/projects';
+import { sortProjects } from 'core/projects';
 
 export const getAllProjects = (state: storeShape): projectType[] => state.projects.allIds
   .map(id => state.projects.byId[id])
